+++ tests/test_streamer.py
@@ -4,7 +4,7 @@ from time import sleep
 
 import pytest
 
-from mpservice.streamer import Streamer, is_exception
+from mpservice.streamer import Stream
 
 
 def test_stream():
@@ -13,37 +13,37 @@ def test_stream():
             for x in [1, 2, 3]:
                 yield x
 
-    assert Streamer(range(4)).collect() == [0, 1, 2, 3]
-    assert list(Streamer(D())) == [1, 2, 3]
-    assert list(Streamer(['a', 'b', 'c'])) == ['a', 'b', 'c']
+    assert Stream(range(4)).collect() == [0, 1, 2, 3]
+    assert list(Stream(D())) == [1, 2, 3]
+    assert list(Stream(['a', 'b', 'c'])) == ['a', 'b', 'c']
 
 
 def test_drain():
-    assert Streamer(range(8)).drain() == 8
+    assert Stream(range(8)).drain() == 8
 
 
 def test_collect():
-    assert Streamer(range(3)).collect() == [0, 1, 2]
+    assert Stream(range(3)).collect() == [0, 1, 2]
 
 
 def test_map():
     def inc(x, shift=1):
         return x + shift
 
-    assert Streamer(range(5)).map(inc, shift=2).collect() == [2, 3, 4, 5, 6]
-    assert Streamer(range(5)).map(lambda x: x * 2).collect() == [0, 2, 4, 6, 8]
+    assert Stream(range(5)).map(inc, shift=2).collect() == [2, 3, 4, 5, 6]
+    assert Stream(range(5)).map(lambda x: x * 2).collect() == [0, 2, 4, 6, 8]
 
 
 def test_filter():
-    assert Streamer(range(7)).filter(lambda n: (n % 2) == 0).collect() == [0, 2, 4, 6]
+    assert Stream(range(7)).filter(lambda n: (n % 2) == 0).collect() == [0, 2, 4, 6]
 
     def odd_or_even(x, even=True):
         if even:
             return (x % 2) == 0
         return (x % 2) != 0
 
-    assert Streamer(range(7)).filter(odd_or_even).collect() == [0, 2, 4, 6]
-    assert Streamer(range(7)).filter(odd_or_even, even=False).collect() == [1, 3, 5]
+    assert Stream(range(7)).filter(odd_or_even).collect() == [0, 2, 4, 6]
+    assert Stream(range(7)).filter(odd_or_even, even=False).collect() == [1, 3, 5]
 
     data = [0, 1, 2, 'a', 4, ValueError(8), 6, 7]
 
@@ -56,7 +56,7 @@ def test_filter():
             self._idx += 1
             return z
 
-    assert Streamer(data).filter(Tail(6)).collect() == [6, 7]
+    assert Stream(data).filter(Tail(6)).collect() == [6, 7]
 
     class Head:
         def __init__(self):
@@ -66,28 +66,28 @@ def test_filter():
             self._idx += 1
             return z
 
-    assert list(Streamer((2, 3, 1, 5, 4, 7)).filter(Head())) == [1, 4]
+    assert list(Stream((2, 3, 1, 5, 4, 7)).filter(Head())) == [1, 4]
 
 
 def test_filter_exceptions():
     exc = [1, ValueError(3), 2, IndexError(4), FileNotFoundError(), 3, KeyboardInterrupt(), 4]
 
-    assert Streamer(exc).filter_exceptions(BaseException).collect() == [1, 2, 3, 4]
+    assert Stream(exc).filter_exceptions(BaseException).collect() == [1, 2, 3, 4]
 
-    assert Streamer(exc).filter_exceptions(BaseException, Exception).collect() == exc[:-2] + [exc[-1]]
+    assert Stream(exc).filter_exceptions(BaseException, Exception).collect() == exc[:-2] + [exc[-1]]
 
     with pytest.raises(IndexError):
-        assert Streamer(exc).filter_exceptions(ValueError).collect() == exc
+        assert Stream(exc).filter_exceptions(ValueError).collect() == exc
 
     with pytest.raises(FileNotFoundError):
-        ss = Streamer(exc)
+        ss = Stream(exc)
         assert ss.filter_exceptions((ValueError, IndexError)).collect() == exc
 
-    ss = Streamer(exc)
+    ss = Stream(exc)
     with pytest.raises(KeyboardInterrupt):
         assert ss.filter_exceptions(Exception, FileNotFoundError).collect() == exc
 
-    assert Streamer(exc).filter_exceptions(BaseException, FileNotFoundError).collect() == [1, 2, exc[4], 3, 4]
+    assert Stream(exc).filter_exceptions(BaseException, FileNotFoundError).collect() == [1, 2, exc[4], 3, 4]
 
 
 def test_peek():
@@ -95,83 +95,102 @@ def test_peek():
 
     data = list(range(10))
 
-    s = Streamer(data)
+    s = Stream(data)
     n = s.peek(interval=3).drain()
     assert n == 10
 
     def foo(x):
         print(x)
 
-    assert Streamer(data).peek(print_func=foo, interval=0.6).drain() == 10
+    assert Stream(data).peek(print_func=foo, interval=0.6).drain() == 10
 
-    Streamer(data).peek_every_nth(4).drain()
+    Stream(data).peek_every_nth(4).drain()
 
     exc = [0, 1, 2, ValueError(100), 4]
     # `peek` does not drop exceptions
-    assert Streamer(exc).peek().drain() == len(exc)
+    assert Stream(exc).peek().drain() == len(exc)
 
 
 def test_head():
     data = [0, 1, 2, 3, 'a', 5]
-    assert list(Streamer(data).head(3)) == data[:3]
-    assert list(Streamer(data).head(30)) == data
+    assert list(Stream(data).head(3)) == data[:3]
+    assert list(Stream(data).head(30)) == data
 
 
 def test_tail():
     data = [0, 1, 2, 3, 'a', 5]
 
-    assert Streamer(data).tail(2).collect() == ['a', 5]
-    assert list(Streamer(data).tail(10)) == data
+    assert Stream(data).tail(2).collect() == ['a', 5]
+    assert list(Stream(data).tail(10)) == data
 
 
 def test_groupby():
     data = ['atlas', 'apple', 'answer', 'bee', 'block', 'away', 'peter', 'question', 'plum', 'please']
-    assert Streamer(data).groupby(lambda x: x[0]).collect() == [['atlas', 'apple', 'answer'], ['bee', 'block'], ['away'], ['peter'], ['question'], ['plum', 'please']]
+    assert Stream(data).groupby(lambda x: x[0]).collect() == [['atlas', 'apple', 'answer'], ['bee', 'block'], ['away'], ['peter'], ['question'], ['plum', 'please']]
 
 
 def test_batch():
-    s = Streamer(range(11))
+    s = Stream(range(11))
     assert list(s.batch(3)) == [
         [0, 1, 2], [3, 4, 5], [6, 7, 8], [9, 10]]
 
     assert list(s) == [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9, 10]]
 
-    s = Streamer(list(range(11)))
+    s = Stream(list(range(11)))
     assert list(s.batch(3).unbatch()) == list(range(11))
 
 
 def test_unbatch():
     data = [[0, 1, 2], [], [3, 4], [], [5, 6, 7]]
-    assert Streamer(data).unbatch().collect() == list(range(8))
+    assert Stream(data).unbatch().collect() == list(range(8))
 
 
 def test_accumulate():
     data = list(range(6))
-    assert Streamer(data).accumulate(lambda x, y: x + y).collect() == [0, 1, 3, 6, 10, 15]
-    assert Streamer(data).accumulate(lambda x, y: x + y, 3).collect() == [3, 4, 6, 9, 13, 18]
+    assert Stream(data).accumulate(lambda x, y: x + y).collect() == [0, 1, 3, 6, 10, 15]
+    assert Stream(data).accumulate(lambda x, y: x + y, 3).collect() == [3, 4, 6, 9, 13, 18]
 
     def add(x, y):
         if y % 2 == 0:
             return x + y
         return x - y
 
-    assert Streamer(data).accumulate(add, -1).collect() == [-1, -2, 0, -3, 1, -4]
+    assert Stream(data).accumulate(add, -1).collect() == [-1, -2, 0, -3, 1, -4]
 
 
 def test_buffer():
-    assert list(Streamer(range(11)).buffer(5)) == list(range(11))
-    assert list(Streamer(range(11)).buffer(20)) == list(range(11))
+    assert list(Stream(range(11)).buffer(5)) == list(range(11))
+    assert list(Stream(range(11)).buffer(20)) == list(range(11))
 
 
 def test_buffer_noop():
     # No action if buffer is not used.
-    x = Streamer(range(1000)).buffer(20)
+    x = Stream(range(1000)).buffer(20)
     assert True
 
 
 def test_buffer_batch():
-    n = Streamer(range(19)).buffer(10).batch(5).unbatch().peek(interval=1).drain()
+    n = Stream(range(19)).buffer(10).batch(5).unbatch().peek(interval=1).drain()
     assert n == 19
+
+
+def test_buffer_break():
+    def make_data():
+        for x in range(100):
+            if x == 88:
+                raise ValueError(x)
+            sleep(random.random() * 0.02)
+            yield x
+
+    x = Stream(make_data()).buffer(23).map(lambda x: x * 2)
+    for v in x:
+        if v > 50:
+            break
+
+    x = Stream(make_data()).buffer(23).map(lambda x: x * 2)
+    with pytest.raises(ValueError):
+        for v in x:
+            pass
 
 
 def test_parmap():
@@ -189,14 +208,14 @@ def test_parmap():
     expected = [v + 3.8 for v in SYNC_INPUT]
 
     # Test the deprecated `transform`.
-    assert Streamer(SYNC_INPUT).transform(f1, concurrency=1, executor='thread').collect() == expected
+    assert Stream(SYNC_INPUT).transform(f1, concurrency=1, executor='thread').collect() == expected
 
-    assert list(Streamer(SYNC_INPUT).parmap(f1, num_workers=10, executor='thread')) == expected
+    assert list(Stream(SYNC_INPUT).parmap(f1, num_workers=10, executor='thread')) == expected
 
-    assert list(Streamer(SYNC_INPUT).parmap(f1, num_workers=20, executor='thread')) == expected
+    assert list(Stream(SYNC_INPUT).parmap(f1, num_workers=20, executor='thread')) == expected
 
     expected = [(v + 3.8) * 2 for v in SYNC_INPUT]
-    assert list(Streamer(SYNC_INPUT).parmap(f1, executor='thread').parmap(f2, executor='thread')) == expected
+    assert list(Stream(SYNC_INPUT).parmap(f1, executor='thread').parmap(f2, executor='thread')) == expected
 
     class MySink:
         def __init__(self):
@@ -207,7 +226,7 @@ def test_parmap():
             self.result += x * 3
 
     mysink = MySink()
-    n = Streamer(SYNC_INPUT).parmap(f1, executor='thread').parmap(mysink, executor='thread').drain()
+    n = Stream(SYNC_INPUT).parmap(f1, executor='thread').parmap(mysink, executor='thread').drain()
     assert n == len(SYNC_INPUT)
 
     got = mysink.result
@@ -220,7 +239,7 @@ def test_parmap_noop():
     def foo(n):
         return range(n)
  
-    x = Streamer(range(1000)).parmap(foo, executor='thread')
+    x = Stream(range(1000)).parmap(foo, executor='thread')
     assert True
 
 
@@ -235,22 +254,22 @@ def test_parmap_with_error():
         return x + 2
 
     with pytest.raises(TypeError):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process, executor='thread', num_workers=2)
         zz = list(s)
         print(zz)
 
-    zz = list(Streamer(corrupt_data()).parmap(process, executor='thread', num_workers=2, return_exceptions=True))
+    zz = list(Stream(corrupt_data()).parmap(process, executor='thread', num_workers=2, return_exceptions=True))
     print(zz)
     assert isinstance(zz[5], TypeError)
 
-    s = Streamer(corrupt_data())
+    s = Stream(corrupt_data())
     s.parmap(process, executor='thread', num_workers=2, return_exceptions=True)
     n = 0
     nexc = 0
     for x in s:
         n += 1
-        if is_exception(x):
+        if isinstance(x, BaseException):
             nexc += 1
     assert n == len(data)
     assert nexc == 1
@@ -272,47 +291,47 @@ def test_chain():
         return x - 2
 
     with pytest.raises(TypeError):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process1, executor='thread', num_workers=2)
         s.drain()
 
     with pytest.raises((ValueError, TypeError)):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process2, executor='thread', num_workers=3)
         s.drain()
 
     with pytest.raises((ValueError, TypeError)):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process1, executor='thread', num_workers=2, return_exceptions=True)
         s.parmap(process2, executor='thread', num_workers=4)
         s.drain()
 
     with pytest.raises(TypeError):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process1, executor='thread', num_workers=2)
         s.parmap(process2, executor='thread', num_workers=4, return_exceptions=True)
         s.drain()
 
-    s = Streamer(corrupt_data())
+    s = Stream(corrupt_data())
     s.parmap(process1, executor='thread', num_workers=2, return_exceptions=True)
     s.parmap(process2, executor='thread', num_workers=4, return_exceptions=True)
     s.drain()
 
     with pytest.raises((TypeError, ValueError)):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process1, executor='thread', num_workers=1) #2)
         s.buffer(3)
         s.parmap(process2, executor='thread', num_workers=1) #3)
         s.drain()
 
     with pytest.raises((ValueError, TypeError)):
-        s = Streamer(corrupt_data())
+        s = Stream(corrupt_data())
         s.parmap(process1, executor='thread', num_workers=2, return_exceptions=True)
         s.buffer(2)
         s.parmap(process2, executor='thread', num_workers=3)
         s.drain()
 
-    zz = list(Streamer(corrupt_data())
+    zz = list(Stream(corrupt_data())
              .parmap(process1, executor='thread', num_workers=2, return_exceptions=True)
              .buffer(3)
              .parmap(process2, executor='thread', num_workers=3, return_exceptions=True)
@@ -321,7 +340,7 @@ def test_chain():
     print('')
     print(zz)
 
-    s = Streamer(corrupt_data())
+    s = Stream(corrupt_data())
     s.parmap(process1, executor='thread', num_workers=2, return_exceptions=True)
     s.filter_exceptions(BaseException)
     s.buffer(3)
@@ -331,12 +350,12 @@ def test_chain():
     assert list(s) == [1, 2, 3, 4, 5, 6]
 
 
-def test_early_stop():
+def test_stream_early_stop():
     def double(x):
         sleep(0.5)
         return x * 2
 
-    s = Streamer(range(300000))
+    s = Stream(range(300000))
     z = s.parmap(double, executor='thread', num_workers=3)
     n = 0
     for x in z:
@@ -353,5 +372,53 @@ def double(x):
 
 def test_parmap_mp():
     SYNC_INPUT = list(range(278))
-    got = Streamer(SYNC_INPUT).parmap(double, executor='process', num_workers=4).collect()
+    got = Stream(SYNC_INPUT).parmap(double, executor='process', num_workers=4).collect()
     assert got == [v * 2 for v in SYNC_INPUT]
+
+
+class Pad:
+    def __init__(self, value: str):
+        self._val = value
+
+    def __call__(self, x):
+        return f"{self._val} {x}"
+
+
+padder: Pad
+# A global to be used in other processes
+
+
+def prepare_pad(value: str):
+    global padder
+    padder = Pad(value)
+
+
+def pad_worker(x):
+    return padder(x)
+
+
+def test_parmap_initializer():
+    data = Stream(range(30)).parmap(
+        pad_worker, executor='process', num_workers=3,
+        executor_initializer=prepare_pad,
+        executor_init_args=('abc', ),
+    )
+    assert data.collect() == [f"abc {x}" for x in range(30)]
+
+
+
+def add_four(x):
+    return x + 4
+
+
+def worker1(n):
+    data = Stream(range(n)).parmap(add_four, executor='thread', num_workers=3)
+    return list(data)
+
+
+def test_parmap_nest():
+    data = Stream([10, 20, 30]).parmap(worker1, executor='process', num_workers=3)
+    assert data.collect() == [
+        [v + 4 for v in range(n)]
+        for n in (10, 20, 30)
+    ]