+++ CHANGELOG.md
@@ -10,10 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Deprecated context manager on `Streamer`. Instead, use the object directly.
+- Deprecated function `mpservice.util.is_exception`.
 
 ### Changed
 
 - `Streamer.peek` parameter `interval`: default changed to 1 from 1000.
+- Class `Streamer` is renamed `Stream`; the old class `Stream` was removed.
 
 ### Added or enhanced